+++ geoLocalize/utils.go
@@ -5,7 +5,24 @@ import (
 	"io/ioutil"
 	"net/http"
 	"time"
+
+	"github.com/dompaw/RumAir/airStations"
 )
+
+//'Pmpro' system stations has lat/long coordinates exposed ! https://pmpro.dacsystem.pl/webapp/data/averages?type=chart&start=1561939200&end=1561949200&vars=27LON
+func getStationCoordinates(station *airStations.AirStation) (latitude float64, longitude float64, err error) {
+	var pmproSystemBaseAPIURL string = "https://pmpro.dacsystem.pl/webapp/data"
+
+	curr, currMinus2h := nowAndMInus2hInUnixTimestamp()
+
+	stationLatitudeURI := pmproSystemBaseAPIURL + fmt.Sprintf("/averages?type=chart&start=%v&end=%v&vars=%s", currMinus2h, curr, station.LatitudeSensor)
+	latitude, err = getLatOrLonFromAPI(stationLatitudeURI)
+
+	stationLongitudeURI := pmproSystemBaseAPIURL + fmt.Sprintf("/averages?type=chart&start=%v&end=%v&vars=%s", currMinus2h, curr, station.LongitudeSensor)
+	longitude, err = getLatOrLonFromAPI(stationLongitudeURI)
+
+	return
+}
 
 func doAPIGet(uri string) (bytesRead []byte, err error) {
 	var netResp *http.Response
